Move our JS code into individual files instead of one big file. Ensure this file is built and we load the built version. Add an argument to our REST endpoint to set the format of returned images. Modify our JS to use the base64 encoded images to avoid CORS issues
