+++ src/js/media-modal/models/image.js
@@ -1,0 +1,10 @@
+/**
+ * Model to hold our image data.
+ */
+const Image = Backbone.Model.extend( {
+	defaults: {
+		url: '',
+	},
+} );
+
+export default Image;
